fixed userWater

+++ controllers/userWater.js
@@ -18,9 +18,10 @@ const getWaterIntake = async (req, res) => {
 const addWaterIntake = async (req, res, next) => {
   try {
     const { value } = req.body;
-    const { _id } = req.user;
-    const user = await User.findOne({ _id });
-    const { _id: owner } = user;
+    // const { _id } = req.user;
+    // const user = await User.findOne({ _id });
+    const { _id: owner } = req.user;
+    
     
     const water = await WaterIntake.findOne({ owner, date: todayDate });
     // console.log(water);
@@ -42,6 +43,18 @@ const addWaterIntake = async (req, res, next) => {
 const deleteByIdWater = async (req, res) => {
   try {
     const { _id } = req.body;
+    const ownerId = req.user._id;
+
+    const waterRecord = await WaterIntake.findOne({ _id });
+
+    if (!waterRecord) {
+      return res.status(404).json({ message: "No water intake record found" });
+    }
+
+    if (String(waterRecord.owner) !== String(ownerId)) {
+      return res.status(403).json({ message: "Unauthorized to delete this record" });
+    }
+  
 
     const result = await WaterIntake.findOneAndDelete({ _id });
